Decouple view container wiring and modularize Telegram send
