+++ api/__init__.py
@@ -12,6 +12,7 @@ from .contracts import (
 from ..app.service.navigator_runtime import MissingAlert
 from ..app.service.navigator_runtime.facade import NavigatorFacade
 from ..bootstrap.navigator import assemble as _bootstrap
+from ..bootstrap.navigator.context import ViewContainerFactory
 
 
 async def assemble(
@@ -22,16 +23,22 @@ async def assemble(
         instrumentation: Iterable[NavigatorRuntimeInstrument] | None = None,
         *,
         missing_alert: MissingAlert | None = None,
+        view_container: ViewContainerFactory | None = None,
 ) -> NavigatorLike:
     """Assemble and return a Navigator facade instance."""
 
+    bootstrap_kwargs = {
+        "missing_alert": missing_alert,
+    }
+    if view_container is not None:
+        bootstrap_kwargs["view_container"] = view_container
     bundle = await _bootstrap(
         event=event,
         state=state,
         ledger=ledger,
         scope=scope,
         instrumentation=instrumentation,
-        missing_alert=missing_alert,
+        **bootstrap_kwargs,
     )
     navigator = NavigatorFacade(bundle.runtime)
     return cast(NavigatorLike, navigator)
