+++ bootstrap/navigator/container.py
@@ -4,10 +4,9 @@ from __future__ import annotations
 from navigator.app.service.navigator_runtime import MissingAlert
 from navigator.core.telemetry import Telemetry
 from navigator.infra.di.container import AppContainer
-from navigator.infra.di.container.telegram import TelegramContainer
 
 from .adapter import LedgerAdapter
-from .context import BootstrapContext
+from .context import BootstrapContext, ViewContainerFactory
 
 
 class ContainerFactory:
@@ -18,19 +17,24 @@ class ContainerFactory:
         telemetry: Telemetry,
         *,
         alert: MissingAlert | None = None,
+        view_container: ViewContainerFactory | None = None,
     ) -> None:
         self._telemetry = telemetry
         self._alert = alert or (lambda scope: "")
+        self._view_container = view_container
 
     def create(self, context: BootstrapContext) -> AppContainer:
         alert = context.missing_alert or self._alert
+        view_container = context.view_container or self._view_container
+        if view_container is None:
+            raise ValueError("view_container must be provided to ContainerFactory")
         return AppContainer(
             event=context.event,
             state=context.state,
             ledger=LedgerAdapter(context.ledger),
             alert=alert,
             telemetry=self._telemetry,
-            view_container=TelegramContainer,
+            view_container=view_container,
         )
