+++ presentation/telegram/assembly.py
@@ -11,7 +11,9 @@ from navigator.api import assemble as assemble_navigator
 from navigator.api.contracts import NavigatorRuntimeInstrument
 from navigator.app.service.navigator_runtime import MissingAlert
 from navigator.core.port.factory import ViewLedger
+from navigator.infra.di.container.telegram import TelegramContainer
 from navigator.presentation.navigator import Navigator
+from navigator.bootstrap.navigator.context import ViewContainerFactory
 
 from .alerts import missing
 from .instrumentation import instrument as default_instrument
@@ -33,12 +35,14 @@ class TelegramNavigatorAssembler:
         *,
         instrumentation: Iterable[NavigatorRuntimeInstrument] | None = None,
         missing_alert: MissingAlert | None = None,
+        view_container: ViewContainerFactory = TelegramContainer,
     ) -> None:
         self._ledger = ledger
         self._instrumentation: Sequence[NavigatorRuntimeInstrument] | None = (
             tuple(instrumentation) if instrumentation is not None else None
         )
         self._missing_alert = missing_alert or missing
+        self._view_container = view_container
 
     async def assemble(self, event: TelegramObject, state: FSMContext) -> Navigator:
         instrumentation = self._instrumentation or (default_instrument,)
@@ -49,6 +53,7 @@ class TelegramNavigatorAssembler:
             scope=outline(event),
             instrumentation=instrumentation,
             missing_alert=self._missing_alert,
+            view_container=self._view_container,
         )
         return navigator
